Initial commit
High voltage power supply I have been engineering for some experiments related to a high energy plasma anomaly I observed as a teenager which I speculate was basically small scale ball-lightning. This project was previously part of a larger hierarchy of projects but I have separated it out from that. Some subsystems are currently not in here including a cross platform JavaScript UI