panic logging seems to be working nicely and putting in place events on
putting in place onOpen and onClosed events on IChannelEvents which is inherited by IChannel and a base class ChannelEvents makes easily available. These events must be dispatched in order from the point of view of the bluetooth stack but without blocking the bluetooth stack so i ended up creating TransientWorker which dispatches on the low priority core (the one we keep for non timing critical operations, the other controls the high speed switching in the circuit and stuff). And transientworker will guarantee the order of the events. The reason we require onOpen and onClosed in thise case (but many uses) is to start up a ping timer among other things.