Two different approaches. shared_ptr and internal memory cleanup vs CleanupBucket
are battling it out they just met logaheads at one point in my code and i think im going with CleanupBucket for a number of reasons for this. Primarily because the message wrapper objects are not really intended to be lingering around in code and passed about from thread to thread.