Going extremely well.
Been implementing error handling for abort and panic and stuff and thats all good now. And been modifying the HighSpeedCore to handle this appropriately. Thats all good from waht i can see. About to make a decision on OtherPeripherals for which the hardware has a Tr and Rs which can either be used for conventional uart style come or i2c coms.