Going great!
Software uart debugged, had some issues but for this particular case a low baud rate is more than fine. Coms from the main controller to the peripheral are not the problem. And those seem to work more reliably anyway. The only coms the other way at present is a handshake which comes from either direction depending which board starts first and that handshake is there to just initialize the state. Though technically it can safely show idle for that.